requirements.txt
tensorflow==2.12.0
streamlit==1.25.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.7.0.72  
h5py==3.8.0